Moved the Struct and Enum into GridManager.cs
