--- Assets/_Scripts/Core/GridSystem/CellType.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-using System.Collections.Generic;
-
-namespace Core.GridSystem
-{
-    public enum CellType
-    {
-        Empty,
-        Ship
-    }
-}
