--- Assets/_Scripts/Core/GridSystem/CellType.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: ccfd9f31e441b8a4f8634bb333bd13f2
