--- Assets/_Scripts/Core/GridSystem/GridPos.cs
@@ -1,20 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using UnityEngine;
-
-namespace Core.GridSystem
-{
-    public struct GridPos
-    {
-        public int x;
-        public int y;
-
-        public GridPos(int x, int y)
-        {
-            this.x = x;
-            this.y = y;
-        }
-    }
-}
