--- Assets/_Scripts/Core/GridSystem/GridPos.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2b0494d22d6c8754ab99ddafd4c54c69
